Changed to apache.commons.math3 random generator

+++ src/nl/wilcotomassen/loremdatum/generator/RandomGeneratorConfiguration.java
@@ -5,7 +5,7 @@ public abstract class RandomGeneratorConfiguration extends GeneratorConfiguratio
 	/**
 	 * Seed for the random number generator
 	 */
-	public long seed;
+	public int seed;
 	
 	/**
 	 * Probability of generating a N/A value [0.0, 1.0f] or
